move Result.Json.Serialization into CSharpFunctionalExtensions.Json.Serialization namespace
